code: rearrange code

+++ src/Modules/Events/Eventify.Modules.Events.Application/TicketTypes/CreateTicketType/CreateTicketTypeCommand.cs
@@ -7,4 +7,4 @@ public sealed record CreateTicketTypeCommand(
     string Name,
     decimal Price,
     string Currency,
-    decimal Quantity) : ICommand<Guid>;
+    int Quantity) : ICommand<Guid>;
